Move cache_res_mgr to shared_cache
